backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Integration
anthropic==0.7.8

# File handling and utilities
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.3

# Google Sheets integration (optional)
gspread==5.12.0
google-auth==2.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production server
gunicorn==21.2.0